Implement deleting users (#5)
* Add deleting users feature

* Improve users list and observe local users from the engine

+++ UserManager/UserManager/Providers/StorageDataProvider.swift
@@ -6,6 +6,7 @@ final class StorageDataProvider {
     enum DataError: Error {
         case fetchFailed
         case savingFailed
+        case deletingFailed
     }
     
     private static let modelName = "DataModel"
@@ -65,9 +66,16 @@ extension StorageDataProvider {
         }
     }
     
-    func deleteUser(_ user: User) {
-        context.delete(user)
-        try? saveContext()
+    func deleteUser(withEmail email: String) throws {
+        do {
+            guard let user = try getUser(byEmail: email) else {
+                throw DataError.fetchFailed
+            }
+            context.delete(user)
+            try saveContext()
+        } catch {
+            throw DataError.deletingFailed
+        }
     }
 }
 
@@ -77,6 +85,18 @@ private extension StorageDataProvider {
             try context.save()
         } catch {
             throw DataError.savingFailed
+        }
+    }
+    
+    func getUser(byEmail email: String) throws -> User? {
+        do {
+            let request = User.fetchRequest() as NSFetchRequest
+            request.predicate = NSPredicate(format: "email == %@", email)
+            request.fetchLimit = 1
+            let user = try context.fetch(request)
+            return user.first
+        } catch {
+            throw DataError.fetchFailed
         }
     }
 }
